+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Old/BILCameraOp.java
@@ -1,6 +1,5 @@
-package org.firstinspires.ftc.teamcode;
+package org.firstinspires.ftc.teamcode.Old;
 
-import org.firstinspires.ftc.teamcode.BILCameraPreview;
 // import org.firstinspires.ftc.robotcontroller.internal.testcode.FtcRobotControllerActivity;
 import com.qualcomm.robotcore.eventloop.opmode.OpMode;
 import java.io.ByteArrayOutputStream;
