+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Old/BILEncoderAutonomousLeft.java
@@ -1,19 +1,12 @@
-package org.firstinspires.ftc.teamcode;
+package org.firstinspires.ftc.teamcode.Old;
 
 
-import com.qualcomm.robotcore.eventloop.opmode.OpMode;
-import com.qualcomm.robotcore.hardware.DcMotor;
-import com.qualcomm.robotcore.hardware.DcMotorController;
 import com.qualcomm.robotcore.hardware.Servo;
-
-import java.util.Map;
-import java.util.Set;
 
 /**
  * Created by nill on 11/19/15.
  */
-public class BILEncoderAutonomousRight extends BILAutonomousCommon2015
-{
+public class BILEncoderAutonomousLeft extends BILAutonomousCommon2015 {
     //DcMotor motorRight;
     //DcMotor motorLeft;
 
@@ -22,14 +15,12 @@ public class BILEncoderAutonomousRight extends BILAutonomousCommon2015
     private int v_state;
 
     @Override
-    public void init()
-    {
-        super.init(); //calls from BILAutonomousCommon
-        resetDriveEncoders(); //resets encoder values back to original
-        v_state = 0; //resets the state of the drive loop
-        /*Set<Map.Entry<String, DcMotorController>> controllerSet = hardwareMap.dcMotorController.entrySet();
-        for (Map.Entry<String, DcMotorController> entry: controllerSet)
-        {
+    public void init() {
+        super.init();
+        resetDriveEncoders();
+        v_state = 0;
+       /* Set<Map.Entry<String, DcMotorController>> controllerSet = hardwareMap.dcMotorController.entrySet();
+        for (Map.Entry<String, DcMotorController> entry : controllerSet) {
             System.out.println("Entry key:" + entry.getKey());
             System.out.println("Entry value:" + entry.getValue());
         }*/
@@ -41,12 +32,10 @@ public class BILEncoderAutonomousRight extends BILAutonomousCommon2015
 	 */
 
     @Override
-    public void loop()
-    {
-        switch(v_state)
-        {
+    public void loop() {
+        switch (v_state) {
             case 0:
-                if (!gyroSensor.isCalibrating()){ //if the gyro sensor is calibrating (flashing blue)
+                if (!gyroSensor.isCalibrating()){
                     v_state++;
                 }
                 break;
@@ -56,80 +45,77 @@ public class BILEncoderAutonomousRight extends BILAutonomousCommon2015
                 setForwardDriveDistance(7.0, circum); //drives backwards for 4.5 feet
                 calculateSpeedBasedOnGyro();
                 driveForward(speedContainer);
-                if (haveDriveEncodersReached() || getDistance() < 8) {
+
+                if (haveDriveEncodersReached() || getDistance() < 8)
+                {
                     stopDriving();
+
                     v_state++;
                 }
                 break;
-            /*case 2:
+           /* case 1:
                 System.out.println("********************** " + v_state);
                 resetDriveEncoders();
-                if (areDriveEncodersZero())
-                {
+                if (areDriveEncodersZero()) {
                     v_state++;
                 }
 
                 break;
-            case 3:
+            case 2:
                 System.out.println("*********** " + v_state);
                 runUsingEncoders();
                 setForwardDriveDistance(2.5, circum);
                 driveForward(); //drive forward 2 feet
 
-                if(haveDriveEncodersReached())
-                {
+                if (haveDriveEncodersReached()) {
                     stopDriving();
                     v_state++;
                 }
+                break;
+            case 3:
+                resetDriveEncoders();
+                System.out.println("********************** " + v_state);
+                resetDriveEncoders();
+                if (areDriveEncodersZero()) {
+                    v_state++;
+                }
+
                 break;
             case 4:
-                resetDriveEncoders();
-                System.out.println("********************** " + v_state);
-                resetDriveEncoders();
-                if (areDriveEncodersZero())
-                {
-                    v_state++;
-                }
-
-                break;
-            case 5:
                 runUsingEncoders();
                 setForwardDriveDistance(3.0, circum);
-                turnRight(); //turn left without moving forward
+                turnLeft(); //turn left without moving forward
 
-                if(hasRightDriveEncoderReached(encoderTicks))
-                {
+                if (hasLeftDriveEncoderReached(encoderTicks)) {
                     stopDriving();
                     v_state++;
                 }
                 break;
-            case 6:
+            case 5:
                 resetDriveEncoders();
                 System.out.println("********************** " + v_state);
                 resetDriveEncoders();
-                if (areDriveEncodersZero())
-                {
+                if (areDriveEncodersZero()) {
                     v_state++;
                 }
 
                 break;
-            case 7:
+            case 6:
                 runUsingEncoders();
                 setForwardDriveDistance(5.0, circum);
                 driveUpRamp(); //drive forward up ramp
 
-                if(haveDriveEncodersReached())
-                {
+                if (haveDriveEncodersReached()) {
                     stopDriving();
                     v_state++;
                 }
-                break;*/
+                break; */
+
         }
         telemetry.addData("State", v_state);
     }
 
-    private void printToLog()
-    {
+    private void printToLog() {
         System.out.println("left motor: " + motorLeft.getCurrentPosition());
         System.out.println("right motor: " + motorRight.getCurrentPosition());
         System.out.println("encoders mode: " + motorLeft.getMode());
