+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Old/BILSimpleForward.java
@@ -1,12 +1,9 @@
-package org.firstinspires.ftc.teamcode;
+package org.firstinspires.ftc.teamcode.Old;
 
 import com.qualcomm.robotcore.eventloop.opmode.OpMode;
 import com.qualcomm.robotcore.hardware.DcMotor;
 import com.qualcomm.robotcore.hardware.DcMotorController;
 import com.qualcomm.robotcore.hardware.Servo;
-
-import java.util.Map;
-import java.util.Set;
 
 /**
  * Created by nill on 1/7/16.
