+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Old/BILVuforiaBlueBeacons.java
@@ -1,10 +1,13 @@
-package org.firstinspires.ftc.teamcode;
+package org.firstinspires.ftc.teamcode.Old;
 
 import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
 
 import org.firstinspires.ftc.robotcore.external.navigation.VuforiaLocalizer;
 import org.firstinspires.ftc.robotcore.external.navigation.VuforiaTrackable;
 import org.firstinspires.ftc.robotcore.external.navigation.VuforiaTrackables;
+import org.firstinspires.ftc.teamcode.BILAutonomousCommon;
+import org.firstinspires.ftc.teamcode.BILRobotHardware;
+import org.firstinspires.ftc.teamcode.BILVuforiaCommon;
 
 import java.util.List;
